chore: move getSplitItems to a shared folder
